Added a Pure ECS example scene for instanced rendering
Ironically enough though, nothing seems to render no matter what I try to do. All the entities seem to be created and they seem to get all the necessary components. The material is marked with GPU instancing and the bootstrap code is based on Unity's own tutorials and demos as well.